+++ jogo_mt_mt_legal/Player.h
@@ -6,11 +6,11 @@ class Player : public Alive
 private:
 
 	static const int MAX_HP;
-	static const int MAX_HORIZONTAL_SPEED;
-	static const int MAX_VERTICAL_SPEED;
-	static const int ACCELARATION;
+	static const float MAX_HORIZONTAL_SPEED;
+	static const float MAX_VERTICAL_SPEED;
+	static const float ACCELARATION;
 	static const int MAX_JUMP_BUFFER;
-	static const int JUMP_STRENGTH;
+	static const float JUMP_STRENGTH;
 	static const int MAX_JUMP_PERIOD;
 	static const int MAX_ATTACK_BUFFER;
 
@@ -26,7 +26,7 @@ private:
 	int attackCounter;
 	bool hasAttacked;
 	bool wasAttackPressed;
-	int friction;
+	float friction;
 public:
 	Player(const bool isPlayer2, const bool ally, const int health);
 	void movement();
